feat: adiciona validações e mensagens de erro nas rotas de manipulação de tarefas

+++ src/routes.js
@@ -49,11 +49,67 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params;
-      const now = new Date();
+
+      const task = database.select("tasks").find(task => task.id === id);
+
+      if (!task) {
+        return res.writeHead(404).end(JSON.stringify({ message: `Task não encontrada` }))
+      }
 
       database.delete("tasks", id);
 
-      return res.writeHead(204).end();
+      return res.writeHead(200).end(JSON.stringify({ message: `Task ${id} excluida com sucesso!`}));
     },
   },
+  {
+    method: "PUT",
+    path: buildRoutePath("/tasks/:id"),
+    handler: (req, res) => {
+      const { id } = req.params;
+      const { title, description } = req.body;
+      const now = new Date();
+
+      if (!title && !description) {
+        return res.writeHead(404).end(JSON.stringify({ message: "É necessário informar pelo menos um dos campos."}))
+      }
+
+      const task = database.select("tasks").find(task => task.id === id);
+
+      if (!task) {
+        return res.writeHead(404).end(JSON.stringify({ message: "Task não encontrada" }));
+      }
+
+      const data = {
+        updated_at: now,
+      };
+    
+      if (title) data.title = title;
+      if (description) data.description = description;
+    
+      database.update("tasks", id, data);
+    
+      return res.writeHead(200).end(JSON.stringify({ message: `Task ${id} atualizada com sucesso!`}));
+    }
+  },
+  {
+    method: 'PATCH',
+    path: buildRoutePath('/tasks/:id/complete'),
+    handler: (req, res) => {
+      const { id } = req.params
+      const now = new Date();
+
+      const task = database.select("tasks").find(task => task.id === id);
+
+      if (!task) {
+        return res.writeHead(404).end(JSON.stringify({ message: "Task não encontrada" }));
+      }
+
+      const isTaskCompleted = !!task.completed_at
+      const completed_at = isTaskCompleted ? null : new Date()
+
+      database.update('tasks', id, { completed_at })
+
+      return res.writeHead(200).end(JSON.stringify({ message: `Task ${id} concluída!`}))
+    }
+  }
 ];
